Split sqlite/indexes.py into multiple files (one per index)
