+++ python/ta/typeagent/storage/sqlite/propindex.py
@@ -1,0 +1,119 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+"""SQLite-based property index implementation."""
+
+import sqlite3
+
+from ...knowpro import interfaces
+from ...knowpro.interfaces import ScoredSemanticRefOrdinal
+
+
+class SqlitePropertyIndex(interfaces.IPropertyToSemanticRefIndex):
+    """SQLite-backed implementation of property to semantic ref index."""
+
+    def __init__(self, db: sqlite3.Connection):
+        self.db = db
+
+    async def size(self) -> int:
+        cursor = self.db.cursor()
+        cursor.execute(
+            "SELECT COUNT(*) FROM (SELECT DISTINCT prop_name, value_str FROM PropertyIndex)"
+        )
+        return cursor.fetchone()[0]
+
+    async def get_values(self) -> list[str]:
+        cursor = self.db.cursor()
+        cursor.execute(
+            "SELECT DISTINCT value_str FROM PropertyIndex ORDER BY value_str"
+        )
+        return [row[0] for row in cursor.fetchall()]
+
+    async def add_property(
+        self,
+        property_name: str,
+        value: str,
+        semantic_ref_ordinal: (
+            interfaces.SemanticRefOrdinal | interfaces.ScoredSemanticRefOrdinal
+        ),
+    ) -> None:
+        # Extract semref_id and score from the ordinal
+        if isinstance(semantic_ref_ordinal, interfaces.ScoredSemanticRefOrdinal):
+            semref_id = semantic_ref_ordinal.semantic_ref_ordinal
+            score = semantic_ref_ordinal.score
+        else:
+            semref_id = semantic_ref_ordinal
+            score = 1.0
+
+        # Normalize property name and value (to match in-memory implementation)
+        from ...storage.memory.propindex import (
+            make_property_term_text,
+            split_property_term_text,
+        )
+
+        term_text = make_property_term_text(property_name, value)
+        term_text = term_text.lower()  # Matches PropertyIndex._prepare_term_text
+        property_name, value = split_property_term_text(term_text)
+        # Remove "prop." prefix that was added by make_property_term_text
+        if property_name.startswith("prop."):
+            property_name = property_name[5:]
+
+        cursor = self.db.cursor()
+        cursor.execute(
+            """
+            INSERT INTO PropertyIndex (prop_name, value_str, score, semref_id)
+            VALUES (?, ?, ?, ?)
+            """,
+            (property_name, value, score, semref_id),
+        )
+
+    async def clear(self) -> None:
+        cursor = self.db.cursor()
+        cursor.execute("DELETE FROM PropertyIndex")
+
+    async def lookup_property(
+        self,
+        property_name: str,
+        value: str,
+    ) -> list[interfaces.ScoredSemanticRefOrdinal] | None:
+        # Normalize property name and value (to match in-memory implementation)
+        from ...storage.memory.propindex import (
+            make_property_term_text,
+            split_property_term_text,
+        )
+
+        term_text = make_property_term_text(property_name, value)
+        term_text = term_text.lower()  # Matches PropertyIndex._prepare_term_text
+        property_name, value = split_property_term_text(term_text)
+        # Remove "prop." prefix that was added by make_property_term_text
+        if property_name.startswith("prop."):
+            property_name = property_name[5:]
+
+        cursor = self.db.cursor()
+        cursor.execute(
+            "SELECT semref_id, score FROM PropertyIndex WHERE prop_name = ? AND value_str = ?",
+            (property_name, value),
+        )
+
+        results = [
+            ScoredSemanticRefOrdinal(semref_id, score)
+            for semref_id, score in cursor.fetchall()
+        ]
+
+        return results if results else None
+
+    async def remove_property(self, prop_name: str, semref_id: int) -> None:
+        """Remove all properties for a specific property name and semantic ref."""
+        cursor = self.db.cursor()
+        cursor.execute(
+            "DELETE FROM PropertyIndex WHERE prop_name = ? AND semref_id = ?",
+            (prop_name, semref_id),
+        )
+
+    async def remove_all_for_semref(self, semref_id: int) -> None:
+        """Remove all properties for a specific semantic ref."""
+        cursor = self.db.cursor()
+        cursor.execute(
+            "DELETE FROM PropertyIndex WHERE semref_id = ?",
+            (semref_id,),
+        )
